BHBC-1958: Move side search bar into features directory
