+++ app/src/features/map/MapPage.tsx
@@ -2,12 +2,12 @@ import Box from '@mui/material/Box';
 import Paper from '@mui/material/Paper';
 import intersect from '@turf/intersect';
 import { IMarkerLayer } from 'components/map/components/MarkerCluster';
-import SideSearchBar from 'components/map/components/SideSearchBar';
 import { IStaticLayer, IStaticLayerFeature } from 'components/map/components/StaticLayers';
 import MapContainer from 'components/map/MapContainer';
 import { AreaToolTip, IFormikAreaUpload } from 'components/upload/UploadArea';
 import { ALL_OF_BC_BOUNDARY, MAP_DEFAULT_ZOOM, SPATIAL_COMPONENT_TYPE } from 'constants/spatial';
 import { IDatasetVisibility, ISearchResult } from 'features/datasets/components/SearchResultList';
+import SideSearchBar from 'features/map/SideSearchBar';
 import { Feature, Polygon } from 'geojson';
 import { useApi } from 'hooks/useApi';
 import useDataLoader from 'hooks/useDataLoader';
