+++ app/src/features/map/SideSearchBar.test.tsx
