+++ app/src/features/map/SideSearchBar.tsx
